🐛fix(architecture): move ArticleHtmlParser to html package + refine ArchUnit serviceNaming (Phase 20 Wave 5 hotfix)
ArchUnit serviceNaming 룰이 ArticleHtmlParser + 4 inner classes 위반 (5건). 1) ArticleHtmlParser 를 service -> html 패키지 이동 (R2-6 SsrfGuard 패턴). 2) serviceNaming 룰에 .arePublic() + .areTopLevelClasses() 추가 - public top-level Service stereotype에만 적용, inner private 구현 세부사항 exempt. 3) ContentExtractService 의 inline 주석 2개 제거 (FileLength 300줄 충족, R4-2/CX-18 의도는 PLAN.md/주변 코드에서 명확).
